requirements.txt
beautifulsoup4==4.11.1
fitz==0.0.1.dev2
openai==0.19.0
python-dotenv==0.20.0
transformers==4.19.2
